+++ src/ts/lib/03-stages/classes/ReleaseStage.ts
@@ -112,6 +112,9 @@ export class ReleaseStage extends AbstractStage<
             '### Changed',
             '- ',
             '',
+            '### Deprecated',
+            '- ',
+            '',
             '### Fixed',
             '- ',
             '',
@@ -598,7 +601,7 @@ export class ReleaseStage extends AbstractStage<
 
         const releaseCmd = `gh release create ${ version } ${ releaseAttachment } ${ this.console.nc.cmdArgs( {
             draft: true,
-            'notes-file': '.releasenotes.md',
+            'notes-file': this.config.paths.notes.release,
             title: `${ version } — ${ timestamp( null, { date: true, time: false } ) }`,
         }, false, false ) }`;